+++ resources/templates/ts/ui.dsl.references.ts
@@ -48,7 +48,9 @@ function createWebviewHtml(): string {
                 })),
               );
             } else {
-              tree.setRootNodes([{ label: 'No references found', removable: false }]);
+              tree.setRootNodes([
+                { label: 'No references found', removable: false, selectable: false },
+              ]);
             }
             macro.progress.hide();
           }